requirements.txt
google-genai
fastapi
uvicorn
pydantic
pyyaml
pandas